# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GiaoDichVien24h API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface UpdateProfileBody {
  name?: string;
  /** @nullable */
  avatar?: string | null;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  MEMBER: "MEMBER",
  GDV: "GDV",
  ADMIN: "ADMIN",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  NORMAL: "NORMAL",
  SCAM: "SCAM",
  TRUSTED: "TRUSTED",
} as const;

export type UserBadge = (typeof UserBadge)[keyof typeof UserBadge];

export const UserBadge = {
  NONE: "NONE",
  TRUSTED_GDV: "TRUSTED_GDV",
  TRUSTED_SELLER: "TRUSTED_SELLER",
} as const;

export interface User {
  id: number;
  /** @nullable */
  uid?: string | null;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: UserRole;
  status: UserStatus;
  badge: UserBadge;
  /** @nullable */
  avatar?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
}

export type SetUserRoleBodyRole =
  (typeof SetUserRoleBodyRole)[keyof typeof SetUserRoleBodyRole];

export const SetUserRoleBodyRole = {
  MEMBER: "MEMBER",
  GDV: "GDV",
  ADMIN: "ADMIN",
} as const;

export interface SetUserRoleBody {
  role: SetUserRoleBodyRole;
}

export type SetUserStatusBodyStatus =
  (typeof SetUserStatusBodyStatus)[keyof typeof SetUserStatusBodyStatus];

export const SetUserStatusBodyStatus = {
  NORMAL: "NORMAL",
  SCAM: "SCAM",
  TRUSTED: "TRUSTED",
} as const;

export type SetUserStatusBodyBadge =
  (typeof SetUserStatusBodyBadge)[keyof typeof SetUserStatusBodyBadge];

export const SetUserStatusBodyBadge = {
  NONE: "NONE",
  TRUSTED_GDV: "TRUSTED_GDV",
  TRUSTED_SELLER: "TRUSTED_SELLER",
} as const;

export interface SetUserStatusBody {
  status?: SetUserStatusBodyStatus;
  badge?: SetUserStatusBodyBadge;
}

export type ScamReportStatus =
  (typeof ScamReportStatus)[keyof typeof ScamReportStatus];

export const ScamReportStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface ScamReport {
  id: number;
  /** @nullable */
  reporterId?: number | null;
  /** @nullable */
  reporterName?: string | null;
  scammerName: string;
  /** @nullable */
  scammerPhone?: string | null;
  /** @nullable */
  scammerBankNumber?: string | null;
  /** @nullable */
  scammerBankName?: string | null;
  /** @nullable */
  scammerSocialLink?: string | null;
  /** @nullable */
  amountLost?: number | null;
  description: string;
  evidenceImages: string[];
  status: ScamReportStatus;
  createdAt: string;
}

export interface Stats {
  totalScammers: number;
  totalAmountLost: number;
  totalReports: number;
  totalMiddlemen: number;
  recentReports: ScamReport[];
}

export interface CreateScamReportBody {
  scammerName: string;
  /** @nullable */
  scammerPhone?: string | null;
  /** @nullable */
  scammerBankNumber?: string | null;
  /** @nullable */
  scammerBankName?: string | null;
  /** @nullable */
  scammerSocialLink?: string | null;
  /** @nullable */
  amountLost?: number | null;
  description: string;
  evidenceImages?: string[];
}

export interface ScamReportListResponse {
  reports: ScamReport[];
  total: number;
  page: number;
  totalPages: number;
}

export interface MiddlemanProfile {
  id: number;
  userId: number;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
  realName: string;
  servicesOffered: string[];
  insuranceFund: number;
  /** @nullable */
  facebookLink?: string | null;
  /** @nullable */
  zaloLink?: string | null;
  /** @nullable */
  telegramLink?: string | null;
  totalTransactions: number;
  successRate: number;
  verificationBadge: boolean;
  createdAt: string;
}

export interface PromoteGdvBody {
  realName: string;
  insuranceFund: number;
  servicesOffered: string[];
  /** @nullable */
  facebookLink?: string | null;
  /** @nullable */
  zaloLink?: string | null;
  /** @nullable */
  telegramLink?: string | null;
}

export type MarketItemStatus =
  (typeof MarketItemStatus)[keyof typeof MarketItemStatus];

export const MarketItemStatus = {
  PENDING: "PENDING",
  AVAILABLE: "AVAILABLE",
  SOLD: "SOLD",
  REJECTED: "REJECTED",
} as const;

export interface MarketItem {
  id: number;
  /** @nullable */
  sellerId?: number | null;
  /** @nullable */
  sellerName?: string | null;
  title: string;
  gameType: string;
  price: number;
  description: string;
  images: string[];
  status: MarketItemStatus;
  createdAt: string;
}

export interface CreateMarketItemBody {
  title: string;
  gameType: string;
  price: number;
  description: string;
  images?: string[];
}

export interface MarketItemListResponse {
  items: MarketItem[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface AdminDashboard {
  pendingReports: number;
  pendingMarketItems: number;
  totalUsers: number;
  totalMiddlemen: number;
  totalMarketItems: number;
  recentActivity: ActivityItem[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListScamReportsParams = {
  search?: string;
  page?: number;
  limit?: number;
};

export type ListMarketItemsParams = {
  gameType?: string;
  page?: number;
  limit?: number;
};

export type AdminListReportsParams = {
  status?: AdminListReportsStatus;
};

export type AdminListReportsStatus =
  (typeof AdminListReportsStatus)[keyof typeof AdminListReportsStatus];

export const AdminListReportsStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export type AdminListMarketItemsParams = {
  status?: AdminListMarketItemsStatus;
};

export type AdminListMarketItemsStatus =
  (typeof AdminListMarketItemsStatus)[keyof typeof AdminListMarketItemsStatus];

export const AdminListMarketItemsStatus = {
  PENDING: "PENDING",
  AVAILABLE: "AVAILABLE",
  SOLD: "SOLD",
  REJECTED: "REJECTED",
} as const;

export type AdminListUsersParams = {
  search?: string;
};