# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GiaoDichVien24h API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/reports:
    get:
      operationId: adminListReports
      tags: [admin]
      summary: Admin list all reports with optional status filter
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScamReport"
  /admin/reports/{id}/approve:
    patch:
      operationId: approveReport
      tags: [admin]
      summary: Approve a scam report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScamReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/reports/{id}/reject:
    patch:
      operationId: rejectReport
      tags: [admin]
      summary: Reject a scam report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScamReport"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/market:
    get:
      operationId: adminListMarketItems
      tags: [admin]
      summary: Admin list all market items (all statuses)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, AVAILABLE, SOLD, REJECTED]
      responses:
        "200":
          description: All market items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketItemListResponse"
  /admin/market/{id}:
    delete:
      operationId: adminDeleteMarketItem
      tags: [admin]
      summary: Admin delete a market item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/market/{id}/approve:
    patch:
      operationId: adminApproveMarketItem
      tags: [admin]
      summary: Admin approve a market item (set to AVAILABLE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/market/{id}/reject:
    patch:
      operationId: adminRejectMarketItem
      tags: [admin]
      summary: Admin reject a market item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: Admin list all users
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}:
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Admin delete a user account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/promote-gdv:
    post:
      operationId: promoteToGdv
      tags: [admin]
      summary: Promote user to GDV (Middleman) role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteGdvBody"
      responses:
        "200":
          description: User promoted to GDV
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MiddlemanProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/role:
    patch:
      operationId: adminSetUserRole
      tags: [admin]
      summary: Admin set user role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserRoleBody"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/status:
    patch:
      operationId: adminSetUserStatus
      tags: [admin]
      summary: Admin set user status and badge
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetUserStatusBody"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/middlemen/{id}:
    delete:
      operationId: adminRemoveGdv
      tags: [admin]
      summary: Admin remove GDV (revoke GDV status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: GDV removed
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        uid:
          type: ["string", "null"]
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [MEMBER, GDV, ADMIN]
        status:
          type: string
          enum: [NORMAL, SCAM, TRUSTED]
        badge:
          type: string
          enum: [NONE, TRUSTED_GDV, TRUSTED_SELLER]
        avatar:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - status
        - badge
        - createdAt
    SetUserRoleBody:
      type: object
      properties:
        role:
          type: string
          enum: [MEMBER, GDV, ADMIN]
      required:
        - role
    SetUserStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [NORMAL, SCAM, TRUSTED]
        badge:
          type: string
          enum: [NONE, TRUSTED_GDV, TRUSTED_SELLER]
...
    ScamReport:
      type: object
      properties:
        id:
          type: integer
        reporterId:
          type: ["integer", "null"]
        reporterName:
          type: ["string", "null"]
        scammerName:
          type: string
        scammerPhone:
          type: ["string", "null"]
        scammerBankNumber:
          type: ["string", "null"]
        scammerBankName:
          type: ["string", "null"]
        scammerSocialLink:
          type: ["string", "null"]
        amountLost:
          type: ["integer", "null"]
        description:
          type: string
        evidenceImages:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        createdAt:
          type: string
      required:
        - id
        - scammerName
        - description
        - evidenceImages
        - status
        - createdAt
...
    MiddlemanProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        realName:
          type: string
        servicesOffered:
          type: array
          items:
            type: string
        insuranceFund:
          type: integer
        facebookLink:
          type: ["string", "null"]
        zaloLink:
          type: ["string", "null"]
        telegramLink:
          type: ["string", "null"]
        totalTransactions:
          type: integer
        successRate:
          type: number
        verificationBadge:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - realName
        - servicesOffered
        - insuranceFund
        - totalTransactions
        - successRate
        - verificationBadge
        - createdAt
    PromoteGdvBody:
      type: object
      properties:
        realName:
          type: string
        insuranceFund:
          type: integer
        servicesOffered:
          type: array
          items:
            type: string
        facebookLink:
          type: ["string", "null"]
        zaloLink:
          type: ["string", "null"]
        telegramLink:
          type: ["string", "null"]
      required:
        - realName
        - insuranceFund
        - servicesOffered
    MarketItem:
      type: object
      properties:
        id:
          type: integer
        sellerId:
          type: ["integer", "null"]
        sellerName:
          type: ["string", "null"]
        title:
          type: string
        gameType:
          type: string
        price:
          type: integer
        description:
          type: string
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [PENDING, AVAILABLE, SOLD, REJECTED]
        createdAt:
          type: string
      required:
        - id
        - title
        - gameType
        - price
        - description
        - images
        - status
        - createdAt
...
    MarketItemListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/MarketItem"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - items
        - total
        - page
        - totalPages
...